Make Eigen blas usage more robust trough -Xclang Definition
